requirements.txt
anyio==3.6.2
black==22.10.0
bleach==5.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
fastapi==0.88.0
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.29
gotrue==0.5.4
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==5.1.0
invoke==1.7.3
jaraco.classes==3.2.3
keyring==23.11.0
mccabe==0.7.0
more-itertools==9.0.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
Pillow==9.3.0
pkginfo==1.9.2
platformdirs==2.5.4
postgrest-py==0.10.3
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.12.0
python-multipart==0.0.5
python-semantic-release==7.32.1
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.3
realtime==0.0.5
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.22.0
storage3==0.3.5
supafunc==0.2.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
zipp==3.11.0
